Extract PortableAnalysisMappers from AnalysisMap, and consume directly in the extractor. Now to actually dump a thing.
