+++ src/scala/org/pantsbuild/zinc/extractor/BUILD
@@ -7,6 +7,7 @@ scala_library(
   ),
   dependencies=[
     '3rdparty/jvm/com/fasterxml/jackson/module:scala',
+    '3rdparty/jvm/org/scala-sbt:zinc',
     'src/scala/org/pantsbuild/zinc/analysis',
     'src/scala/org/pantsbuild/zinc/options',
   ],
