+++ src/scala/org/pantsbuild/zinc/extractor/Main.scala
@@ -1,9 +1,13 @@
 /**
- * Copyright (C) 2012 Typesafe, Inc. <http://www.typesafe.com>
+ * Copyright (C) 2017 Pants project contributors (see CONTRIBUTORS.md).
+ * Licensed under the Apache License, Version 2.0 (see LICENSE).
  */
 
 package org.pantsbuild.zinc.extractor
 
+import sbt.internal.inc.FileBasedStore
+
+import org.pantsbuild.zinc.analysis.PortableAnalysisMappers
 import org.pantsbuild.zinc.options.Parsed
 
 /**
@@ -23,5 +27,18 @@ object Main {
     }
 
     if (settings.help) Settings.printUsage(Command)
+
+    val analysis =
+      FileBasedStore(
+        settings.analysis.cache,
+        new PortableAnalysisMappers(settings.analysis.rebaseMap)
+      )
+        .get()
+        .getOrElse {
+          throw new RuntimeException(s"Failed to load analysis from ${settings.analysis.cache}")
+        }
+        ._1
+
+    println(analysis)
   }
 }
